feat: convert Shw::Autoscrollable
