+++ showcase/app/components/shw/autoscrollable/types.ts
@@ -1,0 +1,7 @@
+export enum AutoscrollableDirectionValues {
+  Both = 'both',
+  X = 'x',
+  Y = 'y',
+}
+
+export type AutoscrollableDirections = `${AutoscrollableDirectionValues}`;
